.Refactor rewrite as test
