+++ test/integration/go.mod
@@ -3,6 +3,8 @@ module github.com/thomasobenaus/sokar/test/integration
 go 1.13
 
 require (
+	github.com/aws/aws-sdk-go v1.25.37
+	github.com/golang/mock v1.4.3
 	github.com/hashicorp/nomad/api v0.0.0-20200505181022-0c86b7a0848c
 	github.com/julienschmidt/httprouter v1.3.0 // indirect
 	github.com/pkg/errors v0.9.1
